fix bug

+++ app/src/main/java/com/vietinterview/getbee/api/response/detailjobcustomer/DetailJobCustomerResponse.java
@@ -63,7 +63,7 @@ public class DetailJobCustomerResponse implements Parcelable {
     private String expireDate;
     @SerializedName("fee")
     @Expose
-    private Integer fee;
+    private Long fee;
     @SerializedName("fromSalary")
     @Expose
     private Integer fromSalary;
@@ -212,7 +212,7 @@ public class DetailJobCustomerResponse implements Parcelable {
         if (in.readByte() == 0) {
             fee = null;
         } else {
-            fee = in.readInt();
+            fee = in.readLong();
         }
         if (in.readByte() == 0) {
             fromSalary = null;
@@ -434,11 +434,11 @@ public class DetailJobCustomerResponse implements Parcelable {
         this.expireDate = expireDate;
     }
 
-    public Integer getFee() {
+    public Long getFee() {
         return fee;
     }
 
-    public void setFee(Integer fee) {
+    public void setFee(Long fee) {
         this.fee = fee;
     }
 
@@ -743,7 +743,7 @@ public class DetailJobCustomerResponse implements Parcelable {
             parcel.writeByte((byte) 0);
         } else {
             parcel.writeByte((byte) 1);
-            parcel.writeInt(fee);
+            parcel.writeLong(fee);
         }
         if (fromSalary == null) {
             parcel.writeByte((byte) 0);
